[Build] Clear warnings

+++ src/consensus/tx_verify.cpp
@@ -427,7 +427,7 @@ bool CheckTransaction(const CTransaction& tx, CValidationState &state, bool fSki
         if (tx.vin[0].scriptSig.size() < 2 || tx.vin[0].scriptSig.size() > 100)
             return state.DoS(100, false, REJECT_INVALID, "bad-cb-length");
     } else if (tx.IsZerocoinSpend()) {
-        if (tx.vin.size() < 1 || static_cast<int>(tx.vin.size()) > Params().Zerocoin_MaxSpendsPerTransaction())
+        if (tx.vin.size() < 1 || tx.vin.size() > Params().Zerocoin_MaxSpendsPerTransaction())
             return state.DoS(10, error("CheckTransaction() : Zerocoin Spend has more than allowed txin's"),
                              REJECT_INVALID, "bad-zerocoinspend");
     } else {
